requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
websockets
asyncio
pydantic
mysql-connector-python
requests
python-dotenv